requirements.txt
httpx==0.27.2
python-dotenv==1.0.1
tenacity==9.0.0
tqdm==4.66.5
pydantic==2.9.2
pytest==8.3.2
